remove utils class and move methods to IEX, allow sandbox mode to be changed from cli args
