+++ jtrader/core/hqm.py
@@ -7,7 +7,6 @@ from statistics import mean
 import pandas as pd
 from scipy import stats
 
-import jtrader.core.utils as utils
 from jtrader.core.iex import IEX
 
 stocks = pd.read_csv('files/sp_500_stocks.csv')
@@ -38,7 +37,7 @@ class HighQualityMomentum(IEX):
     def run(self):
         df = pd.DataFrame(columns=csv_columns)
 
-        symbol_groups = list(utils.chunks(stocks['Ticker'], 100))
+        symbol_groups = list(self.chunks(stocks['Ticker'], 100))
         symbol_strings = []
         for i in range(0, len(symbol_groups)):
             symbol_strings.append(','.join(symbol_groups[i]))
@@ -144,4 +143,4 @@ class HighQualityMomentum(IEX):
         writer.save()
 
         with open(file_name, 'rb') as f:
-            utils.send_slack_file(file_name, 'HighQualityMomentum.xlsx', file=io.BytesIO(f.read()))
+            self.send_slack_file(file_name, 'HighQualityMomentum.xlsx', file=io.BytesIO(f.read()))
