+++ jtrader/core/value.py
@@ -7,7 +7,6 @@ import numpy as np
 import pandas as pd
 from scipy import stats
 
-import jtrader.core.utils as utils
 from jtrader.core.iex import IEX
 
 stocks = pd.read_csv('files/sp_500_stocks.csv')
@@ -41,7 +40,7 @@ class Value(IEX):
     def run(self):
         df = pd.DataFrame(columns=csv_columns)
 
-        symbol_groups = list(utils.chunks(stocks['Ticker'], 100))
+        symbol_groups = list(self.chunks(stocks['Ticker'], 100))
         symbol_strings = []
         for i in range(0, len(symbol_groups)):
             symbol_strings.append(','.join(symbol_groups[i]))
@@ -180,4 +179,4 @@ class Value(IEX):
         writer.save()
 
         with open(file_name, 'rb') as f:
-            utils.send_slack_file(file_name, file_name, file=io.BytesIO(f.read()))
+            self.send_slack_file(file_name, file_name, file=io.BytesIO(f.read()))
